extract TourTitleInput
