+++ resources/camino-creator/router.js
@@ -1,6 +1,6 @@
 import { createRouter, createWebHistory } from "vue-router";
 import HomePage from "./pages/HomePage.vue";
-import EditPage from "./pages/EditPage.vue";
+import EditPage from "./pages/EditPage/EditPage.vue";
 import TourStopPage from "./pages/TourStopPage.vue";
 import FeedbackViewPage from "./pages/FeedbackViewPage.vue";
